stop DeparturesBottomSheet from glitching by reading 2 values from database & restart alarms for devs
