+++ app/src/main/java/edu/rpi/shuttletracker/data/models/Departure.kt
@@ -1,5 +1,6 @@
 package edu.rpi.shuttletracker.data.models
 
+import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
@@ -44,7 +45,7 @@ data class Departure(
         time = LocalDateTime.now().with(LocalTime.of(hour, minute)).toString()
     }
 
-    fun getAlarmIntent(context: Context): List<PendingIntent> =
+    private fun getAlarmIntent(context: Context): List<PendingIntent> =
         days.map { day ->
             val requestCode = id * 7 + (day - 1)
             Intent(context, AlarmReceiver::class.java).apply {
@@ -56,7 +57,7 @@ data class Departure(
             }
         }
 
-    fun getMillis(): List<Long> =
+    private fun getMillis(): List<Long> =
         days.map { day ->
             val calendar =
                 Calendar.getInstance().apply {
@@ -75,4 +76,28 @@ data class Departure(
 
             return@map calendar.timeInMillis
         }
+
+    fun cancelAlarms(context: Context) {
+        val alarmMgr = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+        // removes all the alarms for the id
+        for (intent in getAlarmIntent(context)) {
+            alarmMgr.cancel(intent)
+        }
+    }
+
+    fun initiateAlarms(context: Context) {
+        val alarmMgr = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+        getAlarmIntent(context).zip(
+            getMillis(),
+        ) { intent, millis ->
+            alarmMgr.setRepeating(
+                AlarmManager.RTC_WAKEUP,
+                millis,
+                AlarmManager.INTERVAL_DAY * 7,
+                intent,
+            )
+        }
+    }
 }
