+++ app/src/main/java/edu/rpi/shuttletracker/data/repositories/DeparturesRepository.kt
@@ -1,6 +1,5 @@
 package edu.rpi.shuttletracker.data.repositories
 
-import android.app.AlarmManager
 import android.content.Context
 import dagger.hilt.android.qualifiers.ApplicationContext
 import edu.rpi.shuttletracker.data.database.daos.DepartureDao
@@ -19,36 +18,17 @@ class DeparturesRepository
         fun getDepartures(name: String): Flow<List<Departure>> = departureDao.getDepartures(name)
 
         suspend fun insertDeparture(departure: Departure) {
-            val alarmMgr = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-
             // removes all the alarms for the id
-            for (intent in departureDao.getDeparture(departure.id)?.getAlarmIntent(context) ?: listOf()) {
-                alarmMgr.cancel(intent)
-            }
+            departureDao.getDeparture(departure.id)?.cancelAlarms(context)
 
             // ensures we don't modify the old departure
             val newDeparture = departure.copy(id = departureDao.insertDeparture(departure).toInt())
 
-            newDeparture.getAlarmIntent(context).zip(
-                departure.getMillis(),
-            ) { intent, millis ->
-                alarmMgr.setRepeating(
-                    AlarmManager.RTC_WAKEUP,
-                    millis,
-                    AlarmManager.INTERVAL_DAY * 7,
-                    intent,
-                )
-            }
+            newDeparture.initiateAlarms(context)
         }
 
         suspend fun deleteDeparture(departure: Departure) {
-            val alarmMgr = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-
-            // removes all the alarms for the id
-            for (intent in departure.getAlarmIntent(context)) {
-                alarmMgr.cancel(intent)
-            }
-
+            departure.cancelAlarms(context)
             departureDao.deleteDeparture(departure)
         }
     }
