+++ app/src/main/java/edu/rpi/shuttletracker/ui/maps/DeparturesBottomSheet.kt
@@ -6,7 +6,6 @@ import android.content.Intent
 import android.os.Build
 import android.provider.Settings
 import android.widget.Toast
-import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -64,7 +63,7 @@ private val daysOfWeek =
         "Saturday" to "S",
     )
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun DeparturesBottomSheet(
     stop: Stop?,
