+++ app/src/main/java/edu/rpi/shuttletracker/util/Restarter.kt
@@ -5,6 +5,7 @@ import android.content.Context
 import android.content.Intent
 import dagger.Lazy
 import dagger.hilt.android.AndroidEntryPoint
+import edu.rpi.shuttletracker.data.repositories.DeparturesRepository
 import edu.rpi.shuttletracker.data.repositories.UserPreferencesRepository
 import edu.rpi.shuttletracker.util.services.BeaconService
 import kotlinx.coroutines.Dispatchers
@@ -17,13 +18,22 @@ class Restarter : BroadcastReceiver() {
     @Inject
     lateinit var userPreferencesRepository: Lazy<UserPreferencesRepository>
 
+    @Inject
+    lateinit var departuresRepository: DeparturesRepository
+
     override fun onReceive(
         context: Context,
         intent: Intent,
     ) {
         when (intent.action) {
-            Intent.ACTION_MY_PACKAGE_REPLACED -> startBeaconService(context)
-            Intent.ACTION_BOOT_COMPLETED -> startBeaconService(context)
+            Intent.ACTION_MY_PACKAGE_REPLACED -> {
+                startAllAlarms(context)
+                startBeaconService(context)
+            }
+            Intent.ACTION_BOOT_COMPLETED -> {
+                startAllAlarms(context)
+                startBeaconService(context)
+            }
         }
     }
 
@@ -32,6 +42,12 @@ class Restarter : BroadcastReceiver() {
             if (userPreferencesRepository.get().getAutoBoardService().first()) {
                 context.startForegroundService(Intent(context, BeaconService::class.java))
             }
+        }
+    }
+
+    private fun startAllAlarms(context: Context) {
+        runBlocking(Dispatchers.IO) {
+            departuresRepository.getAllDepartures().first().forEach { it.initiateAlarms(context) }
         }
     }
 }
